Some little bug fixes.
Bugs related to login-register systems have been fixed. Also, users are not allowed to register with the names that contains integers only or the letters which are being planned to be used choice sections(e.g. a, b, x, 1, 123 and etc.). The same "name rule" is applied to the names of quizzes.